requirements.txt
Pillow
matplotlib
opencv-python
numpy
